+++ .rubocop_todo.yml
@@ -5,7 +5,3 @@
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
-
-# Offense count: 8
-RSpec/NestedGroups:
-  Max: 5
